+++ src/components/Profile.js
@@ -2,8 +2,8 @@ import React from "react";
 
 const Profile = () => (
   <div>
-    <h2>prifile page</h2>
-    profile
+    <h2>страница профиль</h2>
+    страница профиля
   </div>
 );
